requirements.txt
indent==0.1
psutil==7.0.0
py-cpuinfo==9.0.0
